🚚 Move configuration, tokens, project, and events view tests into admin
