+++ src/admin/views/__tests__/ConfigurationView.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import wait from 'waait';
 import {render} from '@testing-library/react';
 import {MockedProvider} from '@apollo/react-testing';
-import {mocks} from '../../../__mocks__/kf-api-study-creator/mocks';
+import {mocks} from '../../../../__mocks__/kf-api-study-creator/mocks';
 import ConfigurationView from '../ConfigurationView';
 
 jest.mock('auth0-js');
